+++ test/mpi/f-mpi/bsend_init.F90
@@ -64,7 +64,7 @@ PROGRAM bsend_init
 
    ! allocating and attatching MPI-buffered mode buffer
    ALLOCATE(buffer(nints+MPI_BSEND_OVERHEAD))
-   bufsize = nints*C_SIZEOF(buffer(1))+MPI_BSEND_OVERHEAD
+   bufsize = INT(nints*C_SIZEOF(buffer(1))+MPI_BSEND_OVERHEAD)
    CALL MPI_Buffer_attach(buffer, bufsize, ierr)
 
    ! Message cycle
